word-to-pdf_converter/pyproject.toml
[tool.poetry]
name = "word-to-pdf-converter"
version = "0.1.1"
description = "This is a simple word to pdf converter."
authors = ["Harshit Mishra <harshit707125@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
flask = "^3.0.3"
python-docx = "^1.1.2"
pdfkit = "^1.0.0"
pywin32 = "^306"
reportlab = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
